+++ investment_support_directory/helpers.py
@@ -69,7 +69,7 @@ def get_results_from_search_response(response):
 
 
 def get_paginator_url(filters):
-    url = reverse('investment-support-directory-search')
+    url = reverse('investment-support-directory:search')
     querystring = urlencode({
         key: value
         for key, value in filters.items()