+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Pre-release
+
+### Implemented enhancements
+
+### Fixed bugs
+-- TT-1487 403-on-specialist-filters upgrade directory components
+-- TT-1485 Make ISD root 
+
+
 ## [2019.05.16](https://github.com/uktrade/directory-ui-supplier/releases/tag/2019.05.16)
 
 [Full Changelog](https://github.com/uktrade/directory-ui-supplier/compare/2019.04.14...2019.05.16)